+++ src/Classes/CountriesUtil.php
@@ -15,11 +15,12 @@ declare(strict_types=1);
 
 namespace WEM\UtilsBundle\Classes;
 
+use Contao\StringUtil;
 use Contao\System;
+use Exception;
 
 class CountriesUtil
 {
-
     /**
      *  @return array<mixed>
      */
@@ -28,5 +29,295 @@ class CountriesUtil
         $arrCountries = System::getContainer()->get('contao.intl.countries')->getCountries();
 
         return array_combine(array_map('strtolower', array_keys($arrCountries)), $arrCountries);
+    }
+
+    /**
+     * Try to find an ISO Code from the Country fullname.
+     */
+    public static function getCountryISOCodeFromFullname(
+        $strFullname
+    ) {
+        $arrCountries = self::getCountries();
+
+        foreach ($arrCountries as $strIsoCode => $strName) {
+            // Use Generate Alias to handle little imperfections
+            if (
+
+                StringUtil::generateAlias(
+                    $strName
+                ) === StringUtil::generateAlias(
+                    $strFullname
+                )
+
+            ) {
+                return $strIsoCode;
+            }
+        }
+
+        // If nothing, send an exception, because the name is wrong
+        throw new Exception(\sprintf(
+            $GLOBALS['TL_LANG']['WEM']['LOCATIONS']['ERROR']['countryNotFound'],
+            $strFullname
+        ));
+    }
+
+    /**
+     * Return the Continent ISOCode of a Country.
+     *
+     * @param string $strCountry [Country ISOCode]
+     *
+     * @return string [Continent ISOCode]
+     */
+    public static function getCountryContinent(
+        string $strCountry
+    ): string|null {
+        $COUNTRY_CONTINENTS = [
+            'AF' => 'AS',
+            'AX' => 'EU',
+            'AL' => 'EU',
+            'DZ' => 'AF',
+            'AS' => 'OC',
+            'AD' => 'EU',
+            'AO' => 'AF',
+            'AI' => 'NA',
+            'AQ' => 'AN',
+            'AG' => 'NA',
+            'AR' => 'SA',
+            'AM' => 'AS',
+            'AW' => 'NA',
+            'AU' => 'OC',
+            'AT' => 'EU',
+            'AZ' => 'AS',
+            'BS' => 'NA',
+            'BH' => 'AS',
+            'BD' => 'AS',
+            'BB' => 'NA',
+            'BY' => 'EU',
+            'BE' => 'EU',
+            'BZ' => 'NA',
+            'BJ' => 'AF',
+            'BM' => 'NA',
+            'BT' => 'AS',
+            'BO' => 'SA',
+            'BA' => 'EU',
+            'BW' => 'AF',
+            'BV' => 'AN',
+            'BR' => 'SA',
+            'IO' => 'AS',
+            'BN' => 'AS',
+            'BG' => 'EU',
+            'BF' => 'AF',
+            'BI' => 'AF',
+            'KH' => 'AS',
+            'CM' => 'AF',
+            'CA' => 'NA',
+            'CV' => 'AF',
+            'KY' => 'NA',
+            'CF' => 'AF',
+            'TD' => 'AF',
+            'CL' => 'SA',
+            'CN' => 'AS',
+            'CX' => 'AS',
+            'CC' => 'AS',
+            'CO' => 'SA',
+            'KM' => 'AF',
+            'CD' => 'AF',
+            'CG' => 'AF',
+            'CK' => 'OC',
+            'CR' => 'NA',
+            'CI' => 'AF',
+            'HR' => 'EU',
+            'CU' => 'NA',
+            'CY' => 'AS',
+            'CZ' => 'EU',
+            'DK' => 'EU',
+            'DJ' => 'AF',
+            'DM' => 'NA',
+            'DO' => 'NA',
+            'EC' => 'SA',
+            'EG' => 'AF',
+            'SV' => 'NA',
+            'GQ' => 'AF',
+            'ER' => 'AF',
+            'EE' => 'EU',
+            'ET' => 'AF',
+            'FO' => 'EU',
+            'FK' => 'SA',
+            'FJ' => 'OC',
+            'FI' => 'EU',
+            'FR' => 'EU',
+            'GF' => 'SA',
+            'PF' => 'OC',
+            'TF' => 'AN',
+            'GA' => 'AF',
+            'GM' => 'AF',
+            'GE' => 'AS',
+            'DE' => 'EU',
+            'GH' => 'AF',
+            'GI' => 'EU',
+            'GR' => 'EU',
+            'GL' => 'NA',
+            'GD' => 'NA',
+            'GP' => 'NA',
+            'GU' => 'OC',
+            'GT' => 'NA',
+            'GG' => 'EU',
+            'GN' => 'AF',
+            'GW' => 'AF',
+            'GY' => 'SA',
+            'HT' => 'NA',
+            'HM' => 'AN',
+            'VA' => 'EU',
+            'HN' => 'NA',
+            'HK' => 'AS',
+            'HU' => 'EU',
+            'IS' => 'EU',
+            'IN' => 'AS',
+            'ID' => 'AS',
+            'IR' => 'AS',
+            'IQ' => 'AS',
+            'IE' => 'EU',
+            'IM' => 'EU',
+            'IL' => 'AS',
+            'IT' => 'EU',
+            'JM' => 'NA',
+            'JP' => 'AS',
+            'JE' => 'EU',
+            'JO' => 'AS',
+            'KZ' => 'AS',
+            'KE' => 'AF',
+            'KI' => 'OC',
+            'KP' => 'AS',
+            'KR' => 'AS',
+            'KW' => 'AS',
+            'KG' => 'AS',
+            'LA' => 'AS',
+            'LV' => 'EU',
+            'LB' => 'AS',
+            'LS' => 'AF',
+            'LR' => 'AF',
+            'LY' => 'AF',
+            'LI' => 'EU',
+            'LT' => 'EU',
+            'LU' => 'EU',
+            'MO' => 'AS',
+            'MK' => 'EU',
+            'MG' => 'AF',
+            'MW' => 'AF',
+            'MY' => 'AS',
+            'MV' => 'AS',
+            'ML' => 'AF',
+            'MT' => 'EU',
+            'MH' => 'OC',
+            'MQ' => 'NA',
+            'MR' => 'AF',
+            'MU' => 'AF',
+            'YT' => 'AF',
+            'MX' => 'NA',
+            'FM' => 'OC',
+            'MD' => 'EU',
+            'MC' => 'EU',
+            'MN' => 'AS',
+            'ME' => 'EU',
+            'MS' => 'NA',
+            'MA' => 'AF',
+            'MZ' => 'AF',
+            'MM' => 'AS',
+            'NA' => 'AF',
+            'NR' => 'OC',
+            'NP' => 'AS',
+            'AN' => 'NA',
+            'NL' => 'EU',
+            'NC' => 'OC',
+            'NZ' => 'OC',
+            'NI' => 'NA',
+            'NE' => 'AF',
+            'NG' => 'AF',
+            'NU' => 'OC',
+            'NF' => 'OC',
+            'MP' => 'OC',
+            'NO' => 'EU',
+            'OM' => 'AS',
+            'PK' => 'AS',
+            'PW' => 'OC',
+            'PS' => 'AS',
+            'PA' => 'NA',
+            'PG' => 'OC',
+            'PY' => 'SA',
+            'PE' => 'SA',
+            'PH' => 'AS',
+            'PN' => 'OC',
+            'PL' => 'EU',
+            'PT' => 'EU',
+            'PR' => 'NA',
+            'QA' => 'AS',
+            'RE' => 'AF',
+            'RO' => 'EU',
+            'RU' => 'EU',
+            'RW' => 'AF',
+            'SH' => 'AF',
+            'KN' => 'NA',
+            'LC' => 'NA',
+            'PM' => 'NA',
+            'VC' => 'NA',
+            'WS' => 'OC',
+            'SM' => 'EU',
+            'ST' => 'AF',
+            'SA' => 'AS',
+            'SN' => 'AF',
+            'RS' => 'EU',
+            'SC' => 'AF',
+            'SL' => 'AF',
+            'SG' => 'AS',
+            'SK' => 'EU',
+            'SI' => 'EU',
+            'SB' => 'OC',
+            'SO' => 'AF',
+            'ZA' => 'AF',
+            'GS' => 'AN',
+            'ES' => 'EU',
+            'LK' => 'AS',
+            'SD' => 'AF',
+            'SR' => 'SA',
+            'SJ' => 'EU',
+            'SZ' => 'AF',
+            'SE' => 'EU',
+            'CH' => 'EU',
+            'SY' => 'AS',
+            'TW' => 'AS',
+            'TJ' => 'AS',
+            'TZ' => 'AF',
+            'TH' => 'AS',
+            'TL' => 'AS',
+            'TG' => 'AF',
+            'TK' => 'OC',
+            'TO' => 'OC',
+            'TT' => 'NA',
+            'TN' => 'AF',
+            'TR' => 'AS',
+            'TM' => 'AS',
+            'TC' => 'NA',
+            'TV' => 'OC',
+            'UG' => 'AF',
+            'UA' => 'EU',
+            'AE' => 'AS',
+            'GB' => 'EU',
+            'UM' => 'OC',
+            'US' => 'NA',
+            'UY' => 'SA',
+            'UZ' => 'AS',
+            'VU' => 'OC',
+            'VE' => 'SA',
+            'VN' => 'AS',
+            'VG' => 'NA',
+            'VI' => 'NA',
+            'WF' => 'OC',
+            'EH' => 'AF',
+            'YE' => 'AS',
+            'ZM' => 'AF',
+            'ZW' => 'AF',
+        ];
+
+        return $COUNTRY_CONTINENTS[strtoupper($strCountry)];
     }
 }