+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Extension "Utilities" for Contao Open Source CMS
 2.5.0 - 2026-08-27
 - Feat : Add new `WEM\UtilsBundle\Classes\PackageUtil` class
 - Feat : Add new `WEM\UtilsBundle\Classes\PackageUtil::getVersion()` function that return the package version installed
+- Feat : Add new `WEM\UtilsBundle\Classes\CountriesUtil::getCountryISOCodeFromFullname()` function 
+- Feat : Add new `WEM\UtilsBundle\Classes\CountriesUtil::getCountryContinent()` function 
 
 2.4.0 - 2026-08-24
 - Feat : Add new `WEM\UtilsBundle\Classes\PdfUtil` class